add export button to space settings
